hepatitis-detection/backend/requirements.txt
# Core Framework
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0

# Authentication
PyJWT==2.11.0
werkzeug>=2.3.0

# Database
pymongo==4.5.0

# Data Processing
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Machine Learning Models
joblib==1.3.1
xgboost==2.0.0

# API Documentation
flasgger==0.9.7.1

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.12.0

# HTTP Requests
requests==2.31.0

# Additional
pillow>=9.0.0
